(PC-12228) - Add serializer for /providers/<venue_id>
+ move AllocinePivot model to core.providers.models
