+++ api/src/pcapi/scripts/add_allocine_internal_ids.py
@@ -4,9 +4,9 @@ import logging
 from sqlalchemy.orm.util import aliased
 
 from pcapi.core.offerers.models import Venue
+from pcapi.core.providers.models import AllocinePivot
 from pcapi.core.providers.models import AllocineVenueProvider
 from pcapi.models import db
-from pcapi.models.allocine_pivot import AllocinePivot
 
 
 logger = logging.getLogger(__name__)
